Pasando los archivos a Typescript
