+++ package.json
@@ -6,7 +6,7 @@
   "types": "dist",
   "scripts": {
     "test": "jest --config jest.config.json",
-    "build": "tsc --build -p .",
+    "build": "tsc -p .",
     "clean": "tsc --build --clean"
   },
   "keywords": [
@@ -28,7 +28,7 @@
     "time",
     "date"
   ],
-  "type": "module",
+
   "author": "Donatto Minaya",
   "license": "Apache-2.0",
   "repository": {
