+++ src/main.ts
@@ -1,0 +1,37 @@
+"use strict"
+
+// Algebra
+export { SystemEquation } from './algebra/system_equation'
+export { Equations } from './algebra/equations'
+
+// Arithmetic
+export { BasicOperations } from './arithmetic/basic_operations'
+export { RuleOfThree } from './arithmetic/rule_three'
+export { Percentage } from './arithmetic/percentage'
+export { Fractions } from './arithmetic/fractions'
+export { Base } from './arithmetic/base'
+
+// Chemistry
+export { ChemicalElements } from './chemistry/chemical_elements'
+
+// General
+export { LinearRegression } from './general/linear_regression'
+export { Operations } from './general/operations'
+
+// Trigonometry
+export { AngleMeasurements } from './trigonometry/angle_measurements'
+export { Triangles } from './trigonometry/triangles'
+
+// Physics
+export { Vectors } from './physics/vectors.js'
+
+// Others
+export { TerminalColors } from './other/TerminalColors'
+export { Validator } from './other/Validator'
+export { History } from './other/History'
+export { Config } from './other/Config'
+export { Clock } from './other/Clock'
+
+// Browser
+export { BrowserUI } from './browser-ui/browser_ui';
+export { Browser }  from './other/Browser'
