+++ tsconfig.json
@@ -1,11 +1,11 @@
 {
-  "include": ["src/**/*.js"],
+  "include": ["src/**/*.ts"],
   "exclude": ["node_modules"],
   "compilerOptions": {
     "target": "es6",
     "module": "commonjs",
     "strict": true,
-    "allowJs": true,
+
     "declaration": true,
     "declarationMap": true,
     "sourceMap": true,
